precision fix

+++ openmpsrc/src/OpenMPSol.h
@@ -1,4 +1,4 @@
 #pragma once
 /* Function declarations only in header */
-void matmul(const double *A, const double *B, double *C, int n, int k);
-void matmul_ref(const double *A, const double *B, double *C_ref, int n, int k);
+void matmul(const float *A, const float *B, float *C, int n, int k);
+void matmul_ref(const float *A, const float *B, float *C_ref, int n, int k);
